20220530 - Rebase.
20220530 - Clean up.

20220530 - .Net6 Upgrade
