+++ src/Bitcoin/Helpers/Base58.cs
