+++ src/Bitcoin/Helpers/CurveP.cs
