+++ src/Bitcoin/Helpers/Hash.cs
