+++ src/Bitcoin/Helpers/Hex.cs
