+++ src/Bitcoin/Helpers/SubArray.cs
