+++ src/Bitcoin/Keys/PrivateKey.cs
