+++ src/Bitcoin/Keys/PublicKey.cs
@@ -15,15 +15,15 @@ namespace Wallet
 {
     public class PublicKey
     {
+        private static SecureRandom secureRandom = new SecureRandom();
         private static X9ECParameters curve = SecNamedCurves.GetByName("secp256k1");
         private static ECDomainParameters domain = new ECDomainParameters(curve.Curve, curve.G, curve.N, curve.H);
-        private static SecureRandom secureRandom = new SecureRandom();
         
         public static async Task<string> GeneratePublicKey(string privateKey, bool compressed)
         {
             return await Task.Run(() =>
             {
-                BigInteger d = new BigInteger(Hex.toByteArray(privateKey));
+                var d = new BigInteger(Hex.toByteArray(privateKey));
                 Org.BouncyCastle.Math.EC.ECPoint q = domain.G.Multiply(d);
 
                 var publicParams = new ECPublicKeyParameters(q, domain);
@@ -45,14 +45,14 @@ namespace Wallet
         {
             return await Task.Run(() =>
             {
-                BigInteger d = new BigInteger(Hex.toByteArray(privateKey));
+                var d = new BigInteger(Hex.toByteArray(privateKey));
                 Org.BouncyCastle.Math.EC.ECPoint q = domain.G.Multiply(d);
 
                 var publicParams = new ECPublicKeyParameters(q, domain);
                 var encoded = publicParams.Q.GetEncoded();
 
-                BigInteger x = publicParams.Q.XCoord.ToBigInteger();
-                BigInteger y = publicParams.Q.YCoord.ToBigInteger();
+                var x = publicParams.Q.XCoord.ToBigInteger();
+                var y = publicParams.Q.YCoord.ToBigInteger();
 
                 return new Tuple<BigInteger, BigInteger>(x, y);
             });
@@ -65,7 +65,7 @@ namespace Wallet
                 var SHA256 = new SHA256Managed().ComputeHash(Hex.toByteArray(publicKey));
                 var SHA256str = Hex.fromByteArray(SHA256.ToList());
 
-                var RIPEMD160 = new RIPEMD160Managed().ComputeHash(SHA256);
+                var RIPEMD160 = new Crypto.RIPEMD.RIPEMD160Managed().ComputeHash(SHA256);
                 var RIPEMD160str = $"00{ Hex.fromByteArray(RIPEMD160.ToList()) }";
 
                 var version = Hex.toByteArray(RIPEMD160str);
@@ -90,11 +90,11 @@ namespace Wallet
 
         public static bool ValidateBitcoinAddress(string address)
         {
-            if (address.Length < 26 || address.Length > 35) throw new System.Exception("wrong length");
+            if (address.Length < 26 || address.Length > 35) throw new Exception("wrong length");
             var decoded = Base58.Decode(address);
             var d1 = new SHA256Managed().ComputeHash(decoded.SubArray(0, 21));
             var d2 = new SHA256Managed().ComputeHash(d1);
-            if (!decoded.SubArray(21, 4).SequenceEqual(d2.SubArray(0, 4))) throw new System.Exception("bad digest");
+            if (!decoded.SubArray(21, 4).SequenceEqual(d2.SubArray(0, 4))) throw new Exception("bad digest");
             return true;
         }
     }
