+++ src/Bitcoin/Wallet.csproj
