+++ src/Ripper.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30711.63
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ripper", "Ripper\Ripper.csproj", "{E68FB413-377E-4C75-B1AC-633F65DA123D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ripper", "Ripper\Ripper.csproj", "{BE206C45-A4EB-469C-870C-9CA8457F6456}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wallet", "Wallet\Wallet.csproj", "{10EC5397-AA7E-4B93-A89C-725ED9029DD8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bitcoin", "Bitcoin\Bitcoin.csproj", "{43105B08-EBD6-419C-8A4A-2E4195131809}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E68FB413-377E-4C75-B1AC-633F65DA123D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E68FB413-377E-4C75-B1AC-633F65DA123D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E68FB413-377E-4C75-B1AC-633F65DA123D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E68FB413-377E-4C75-B1AC-633F65DA123D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{10EC5397-AA7E-4B93-A89C-725ED9029DD8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{10EC5397-AA7E-4B93-A89C-725ED9029DD8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{10EC5397-AA7E-4B93-A89C-725ED9029DD8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{10EC5397-AA7E-4B93-A89C-725ED9029DD8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BE206C45-A4EB-469C-870C-9CA8457F6456}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BE206C45-A4EB-469C-870C-9CA8457F6456}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BE206C45-A4EB-469C-870C-9CA8457F6456}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BE206C45-A4EB-469C-870C-9CA8457F6456}.Release|Any CPU.Build.0 = Release|Any CPU
+		{43105B08-EBD6-419C-8A4A-2E4195131809}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{43105B08-EBD6-419C-8A4A-2E4195131809}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{43105B08-EBD6-419C-8A4A-2E4195131809}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{43105B08-EBD6-419C-8A4A-2E4195131809}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
